//Mahesh

+++ src/main/java/com/ats/webapi/model/rawmaterial/GetItemDetail.java
@@ -37,7 +37,7 @@ public class GetItemDetail {
 		private int rmUomId;
 		
 		@Column(name="rm_qty")
-		private int rmQty;
+		private float rmQty;
 		
 		@Column(name="rm_weight")
 		private int rmWeight;
@@ -87,11 +87,11 @@ public class GetItemDetail {
 			this.rmType = rmType;
 		}
 
-		public int getRmQty() {
+		public float getRmQty() {
 			return rmQty;
 		}
 
-		public void setRmQty(int rmQty) {
+		public void setRmQty(float rmQty) {
 			this.rmQty = rmQty;
 		}
